+++ js/new_job.js
@@ -38,6 +38,7 @@ function New_Job_GUI() {
         this.categories.load();
     }
 
+/////// CATEGORIES
     this.set_categories_list = function() {
         var option_el = this.$category.children().eq(0).detach();
         for ( var i = 0 ; i < this.categories.categories.length ; i++ ) {
@@ -49,28 +50,69 @@ function New_Job_GUI() {
             this.$category.append(new_option);
         }
     }
+////// end CATEGORIES
 
+
+
+// ISSUES
     this.set_issues_list = function() {
         var option_el = this.$issue.children().eq(0).detach();
 
         if ( this.issues.categories == undefined ) {
-            this.$issue.parent().parent().fadeOut();
             this.$issue.children().remove();
-            this.$issue.append(option_el);
+            new_option = $(option_el).clone();
+            $(new_option).attr("value", " ");
+            $(new_option).attr("title", " ");
+            $(new_option).text("");
+            this.$issue.append(new_option);
+            this.set_issues_status('disabled');
             return;
         };
 
-        this.$issue.parent().parent().fadeIn();
+        this.set_issues_status('enabled');
         this.$issue.children().remove();
         for ( var i = 0 ; i < this.issues.categories.length ; i++ ) {
             issue = this.issues.categories[i];
             new_option = $(option_el).clone();
             $(new_option).attr("value", issue.id);
-            new_option.attr("title", issue.description);
-            new_option.text(issue.name);
+            $(new_option).attr("title", issue.description);
+            $(new_option).text(issue.name);
             this.$issue.append(new_option);
         }
     }
+
+    this.update_issues = function() {
+        this.set_issues_status("load");
+        this.issues.parent_id = this.$category.val();
+        this.issues.load();
+    }
+
+    this.set_issues_status = function(status) {
+        if ( status === 'disabled' ) {
+            if ( this.$issue.attr("disabled") == undefined )
+                this.$issue.attr("disabled", "disabled");
+            this.$issue_load.fadeOut();
+        }
+        else
+        if ( status === 'enabled' ) {
+            if ( this.$issue.attr("disabled") == "disabled" )
+                this.$issue.removeAttr("disabled");
+            JOM['new_job'].$issue_load.fadeOut();
+        }
+        else
+        if ( status === 'start' ) {
+            this.$issue.attr("disabled", "disabled");
+            this.$issue_load.hide();
+        }
+        else
+        if ( status === 'load' ) {
+             if ( this.$issue.attr("disabled") == undefined )
+                this.$issue.attr("disabled", "disabled");
+            this.$issue_load.fadeIn();
+        }
+    }
+////// end ISSUES
+
 
     this.init_events = function() {
         // CLEAR BUTTON
@@ -80,14 +122,9 @@ function New_Job_GUI() {
         // SELECT CATEGORY
         this.$category.unbind().on('change', function(){
             if ( JOM['new_job'].issues.categories != JOM['new_job'].$category.val() ) {
-                JOM['new_job'].update_issue();
+                JOM['new_job'].update_issues();
             }
         });
-    }
-
-    this.update_issue = function() {
-        this.issues.parent_id = this.$category.val();
-        this.issues.load();
     }
 
 
@@ -100,6 +137,7 @@ function New_Job_GUI() {
         this.$issue         = $("#form_new_job [name='issue']");
         this.$assign_to_me  = $("#form_new_job [name='assign_to_me']");
         this.$open_details  = $("#form_new_job [name='open_details']");
+        this.$issue_load    = $("#form_new_job [name='issue']").next();
 
         this.categories       = new Categories();
         this.categories.level = 1;